En byte

+++ AceptaElReto/P116HolaMundo.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class P116HolaMundo {
     public static void main(String[] args) {
         Scanner lector = new Scanner(System.in);
-        int max = lector.nextInt();
-        for(int i = 0; i < max; i++) System.out.println("Hola mundo.");
+        byte max = lector.nextByte();
+        for(byte i = 0; i < max; i++) System.out.println("Hola mundo.");
     }
 }
